make LoadingPage style customizable
