+++ app/components/loading/Loading.js
@@ -3,24 +3,35 @@ import React, { Component } from 'react';
 import type { Node } from 'react';
 import SvgInline from 'react-svg-inline';
 import { observer } from 'mobx-react';
-import { intlShape } from 'react-intl';
+import { defineMessages, intlShape, FormattedMessage } from 'react-intl';
 import classNames from 'classnames';
 import LoadingSpinner from '../widgets/LoadingSpinner';
 import yoroiLogo from '../../assets/images/yoroi-logo-shape-white.inline.svg';
 import styles from './Loading.scss';
-import type { MessageDescriptor } from 'react-intl';
 import environment from '../../environment';
 import LocalizableError from '../../i18n/LocalizableError';
+import globalMessages from '../../i18n/global-messages';
+
+const messages = defineMessages({
+  loading: {
+    id: 'loading.screen.loading',
+    defaultMessage: '!!!loading components',
+  },
+  error: {
+    id: 'loading.screen.error',
+    defaultMessage: '!!!For more help, you can {supportRequestLink}',
+  },
+});
 
 type Props = {|
   currencyIcon: string,
   apiIcon: string,
   isLoadingDataForNextScreen: boolean,
-  loadingDataForNextScreenMessage: MessageDescriptor,
   hasLoadedCurrentLocale: boolean,
   hasLoadedCurrentTheme: boolean,
   error: ?LocalizableError,
-  getErrorMessage: void => Node,
+  onExternalLinkClick: Function,
+  downloadLogs: Function
 |};
 
 @observer
@@ -36,7 +47,6 @@ export default class Loading extends Component<Props> {
       currencyIcon,
       apiIcon,
       isLoadingDataForNextScreen,
-      loadingDataForNextScreenMessage,
       hasLoadedCurrentLocale,
       hasLoadedCurrentTheme,
       error
@@ -72,7 +82,7 @@ export default class Loading extends Component<Props> {
             {isLoadingDataForNextScreen && (
               <div className={styles.loading}>
                 <h1 className={styles.headline}>
-                  {intl.formatMessage(loadingDataForNextScreenMessage)}
+                  {intl.formatMessage(messages.loading)}
                 </h1>
                 <LoadingSpinner />
               </div>
@@ -81,7 +91,7 @@ export default class Loading extends Component<Props> {
               <div className={styles.loading}>
                 <h1 className={styles.error}>
                   {intl.formatMessage(error)}<br />
-                  {this.props.getErrorMessage()}
+                  {this._getErrorMessageComponent()}
                 </h1>
               </div>
             )}
@@ -90,4 +100,39 @@ export default class Loading extends Component<Props> {
       </div>
     );
   }
+
+  _getErrorMessageComponent = (): Node => {
+    const { intl } = this.context;
+    const {
+      onExternalLinkClick,
+      downloadLogs
+    } = this.props;
+
+    const downloadLogsLink = (
+      // eslint-disable-next-line jsx-a11y/anchor-is-valid
+      <a
+        href="#"
+        onClick={_event => downloadLogs()}
+      >
+        {intl.formatMessage(globalMessages.downloadLogsLink)}
+      </a>
+    );
+
+    const supportRequestLink = (
+      <a
+        href={intl.formatMessage(globalMessages.supportRequestLinkUrl)}
+        onClick={event => onExternalLinkClick(event)}
+      >
+        {intl.formatMessage(globalMessages.contactSupport)}
+      </a>
+    );
+
+    return (
+      <p>
+        <FormattedMessage {...globalMessages.logsContent} values={{ downloadLogsLink }} /><br />
+        <FormattedMessage {...messages.error} values={{ supportRequestLink }} />
+      </p>
+    );
+  };
+
 }
